requirements.txt
# Core framework
streamlit>=1.28.0

# Database drivers
psycopg2-binary>=2.9.6
pymongo>=4.6.0
motor>=3.3.2
psutil>=5.9.0

# Image processing
Pillow>=10.0.0

# Web scraping and automation
hyperbrowser
selenium>=4.15.2
beautifulsoup4>=4.12.2
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Utility
watchdog

# Additional utilities for media handling
python-magic>=0.4.27
